% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_dtfr.R
\name{out_dtfr}
\alias{out_dtfr}
\title{Output from list to dataframe}
\usage{
out_dtfr(data)
}
\arguments{
\item{data}{output list object from redFD}
}
\value{
A named dataframe with metrics of FD as columns. "Spnr" is the number of species that was eventually removed from the community.
}
\description{
Output from list to dataframe
}
