% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redFD_null.R
\name{redFD_null}
\alias{redFD_null}
\title{Create simulated communities for null models, remove species and calculate functional diversity indices (based on FD package)}
\usage{
redFD_null(cmnt, trait, spn, rep, permat_args = list(times = 99), ...)
}
\arguments{
\item{cmnt}{matrix containing the abundances of the species in x (or presence-absence, i.e. 0 or 1). Rows are sites and species are columns. Can be missing, in which case dbFD assumes that there is only one community with equal abundances of all species. NAs will be replaced by 0. The number of species (columns) in a must match the number of species (rows) in x. In addition, the species labels in a and x must be identical and in the same order.}

\item{trait}{matrix or data frame of functional traits. Traits can be numeric, ordered, or factor. Binary traits should be numeric and only contain 0 and 1. character traits will be converted to factor. NAs are tolerated.x can also be a species-by-species distance matrix of class dist, in which case NAs are not allowed. When there is only one trait, x can be also be a numeric vector, an ordered factor, or a unordered factor. In all cases, species labels are required.}

\item{spn}{positive integer or vector of positive integers specifying the number or numbers of species to be removed from the community. If spn = 1, all species are sequentially removed. If spn >1, random samples (without replacement) of species equal to spn are removed.}

\item{rep}{positive integer specifying the number of repetitions to remove spn random species from the community.}

\item{permat_args}{list, arguments to be passed to \code{\link[vegan]{permatfull}}.}

\item{...}{arguments to be passed to \code{\link[FD]{dbFD}}.}
}
\value{
A named list with the indices calculated and the simulated community used. The names of each element are the species names that were removed from the community. More species than initially specified, may be removed if they had no appearance in the community. Rows with less than 3 species are removed from the result with a warning. If a community has 0 rows, it is removed with a warning.
}
\description{
Create simulated communities for null models, remove species and calculate functional diversity indices (based on FD package)
}
